Add checkpoint configuration to evolution verbs
